+++ src/Admin/UserList.jsx
@@ -1,9 +1,11 @@
-import { ProfileCard, Title } from "../components/common/Design";
+import { ProfileCard } from "../components/common/Design";
 import { MdOutlineDeleteOutline } from "react-icons/md";
 import { useDispatch, useSelector } from "react-redux";
 import { deleteUserByAdmin, getAllUser } from "../redux/features/authSlice";
 import { useEffect } from "react";
 import { formatDate } from "../utils/formateDate";
+import DashboardTitle from "../components/common/DashboardTitle";
+import Swal from "sweetalert2";
 
 const UserList = () => {
   const dispatch = useDispatch();
@@ -14,17 +16,27 @@ const UserList = () => {
     dispatch(getAllUser());
   }, [dispatch]);
 
-  const handleDeleteUser = async (id) => {
-    await dispatch(deleteUserByAdmin(id));
-    await dispatch(getAllUser());
-  };
 
+  const handleDeleteUser = async (id) => {
+    const result = await Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    });
+  
+    if (result.isConfirmed) {
+      await dispatch(deleteUserByAdmin(id));
+      await dispatch(getAllUser());
+    }
+  };
   return (
     <section className="shadow-s1 p-8 rounded-lg">
       <div className="flex justify-between">
-        <Title level={5} className=" font-normal">
-          User Lists
-        </Title>
+        <DashboardTitle title="User Lists" />
       </div>
       <hr className="my-5" />
       <div className="relative overflow-x-auto rounded-lg">
@@ -75,7 +87,7 @@ const UserList = () => {
                   <td className="px-6 py-4 capitalize">{user?.role}</td>
                   <td className="px-6 py-4">{formatDate(user?.createdAt)}</td>
                   {
-                    user?.role !== "admin" && <td className="py-4 flex justify-end px-8">
+                    user?.role !== "admin" && <td className="py-4 flex justify-end px-8 mt-4 lg:mt-2">
                     <button
                       onClick={() => handleDeleteUser(user?._id)}
                       className="font-medium text-red-500"
